requirements.txt
flask==2.3.2
python-dotenv==1.0.0
click==8.1.3
gunicorn==20.0.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Werkzeug==2.3.4
sqlalchemy==2.0.16
numpy==1.23.5
pandas==2.0.0
psycopg2-binary==2.9.6
pymongo==4.3.3
scikit-learn==1.2.2
scikit-surprise==1.1.3
